+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/BackupInputData.kt
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved
+
+package com.microsoft.did.sdk.datasource.file.models
+
+abstract class BackupInputData {
+    abstract val unprotectedBackup: UnprotectedBackup
+}
