+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/Microsoft2020Backup.kt
@@ -4,7 +4,7 @@ package com.microsoft.did.sdk.datasource.file.models
 
 import com.microsoft.did.sdk.credential.models.VerifiableCredential
 
-class MicrosoftBackup2020(
+class Microsoft2020Backup(
     val walletMetadata: WalletMetadata,
     val verifiableCredentials: List<Pair<VerifiableCredential, VcMetadata>>
 ) : UnprotectedBackup()
