+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/PasswordBackupInputData.kt
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved
+
+package com.microsoft.did.sdk.datasource.file.models
+
+class PasswordBackupInputData(
+    val password: String,
+    override val unprotectedBackup: UnprotectedBackup
+) : BackupInputData()
