--- sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/PasswordProtectedBackupData.kt
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved
-
-package com.microsoft.did.sdk.datasource.file.models
-
-class PasswordProtectedBackupData(
-    val password: String,
-    override val unprotectedBackup: UnprotectedBackup
-) : ProtectedBackupData()
