--- sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/ProtectedBackupData.kt
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved
-
-package com.microsoft.did.sdk.datasource.file.models
-
-abstract class ProtectedBackupData {
-    abstract val unprotectedBackup: UnprotectedBackup
-}
